Update patch set 1
Patch Set 1:

Hello Ryota,

We discussed that some weeks ago. I think the general opinion in the group was to keep the "combined" interface in Promise. The user of that interface can decide to separate resources per operation, if he/she wants to do it as per IFA.

BR,
Arturo

Patch-set: 1
Label: Verified=0