Update patch set 7
Patch Set 7:

> (3 comments)

comment 1: verified on fuel/Danube/virtual pod

comment 2: not sure to catch you here. this part was there in previous versions. I kept this but it would make sense to remove it. It can be managed through config parameters. It is more a question for promise team..up to them to keep it or remove it.
from Functest it will not be used in CI

comment 3: OK I see the point
it is also something that should be fixed by promise team...
I just initiated this in order to ensure that promise would be runnable as it was in Colorado from Functest..the other option will be to support keystone v3 :)

Patch-set: 7
Reviewer: Gerrit User 335 <335@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0