Fix image name iteration
Create a promise image depending on the glance image path.
Due to glance image listing returning image names, looking
for an image based on an image path is always false, thus
its relevant to only look for the glance image name and if it exists
to use its ID, if not to create a new image given the promise
image name.

Change-Id: I2d1347807335f2c7e9d62f1a6c75fb48de6cf448
Signed-off-by: Catalina Focsa <catalina.focsa@enea.com>