Update patch set 5
Patch Set 5:

If there are committers that are listen in the INFO file and not the INFO.yaml, then we have committers that have never accepted their invitation, they are probably not contributing and should be removed. IF they still wish to contribute they need to email helpdesk asking for a new invitation.

Patch-set: 5